refactor(tasks): moved label service to hook
