+++ lib/constants/tasks.ts
@@ -1,0 +1,22 @@
+import type { Priority } from "@/types/enums/notif";
+
+export const PRIORITY_OPTIONS: {
+  value: Priority;
+  label: string;
+  color: string;
+}[] = [
+  { value: "high", label: "High", color: "bg-red-100 text-red-800" },
+  { value: "medium", label: "Medium", color: "bg-yellow-100 text-yellow-800" },
+  { value: "low", label: "Low", color: "bg-green-100 text-green-800" },
+];
+
+export const LABEL_COLORS = [
+  "#ef4444",
+  "#f97316",
+  "#eab308",
+  "#22c55e",
+  "#06b6d4",
+  "#3b82f6",
+  "#8b5cf6",
+  "#ec4899",
+];
